+++ app/src/main/java/com/zekart/tracken/ui/listeners/GasStationAdapterListener.kt
@@ -1,5 +1,5 @@
 package com.zekart.tracken.ui.listeners
 
 interface GasStationAdapterListener{
-    fun onGasStationClick(id_station: Int?)
+    fun onGasStationClick(id_station: Long?)
 }
