+++ app/src/main/java/com/zekart/tracken/viewmodel/StationActivityFactory.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
-class StationActivityFactory (val app: Application, private val id:Int?) : ViewModelProvider.AndroidViewModelFactory(app) {
+class StationActivityFactory (val app: Application, private val id:Long?) : ViewModelProvider.AndroidViewModelFactory(app) {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(ActivityGasStationViewModel::class.java)) {
             @Suppress("UNCHECKED_CAST")
